Chart Display: Refactoring: simplify class names.
