+++ src/main/java/net/imagej/ui/swing/viewer/plot/jfreechart/CategoryChartConverter.java
@@ -11,11 +11,11 @@ import org.scijava.plugin.Plugin;
  * @author Matthias Arzt
  */
 @Plugin(type = Converter.class, priority = Priority.NORMAL_PRIORITY)
-public class CategoryChartToJfcConverter extends AbstractConverter<CategoryChart, JFreeChart> {
+public class CategoryChartConverter extends AbstractConverter<CategoryChart, JFreeChart> {
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T> T convert(Object o, Class<T> aClass) {
-		return (T) new JfcCategoryChartGenerator((CategoryChart) o).getJFreeChart();
+		return (T) new CategoryChartGenerator((CategoryChart) o).getJFreeChart();
 	}
 
 	@Override
