+++ src/main/java/net/imagej/ui/swing/viewer/plot/jfreechart/RendererModifier.java
@@ -3,10 +3,7 @@ package net.imagej.ui.swing.viewer.plot.jfreechart;
 import net.imagej.plot.LineStyle;
 import net.imagej.plot.MarkerStyle;
 import net.imagej.plot.SeriesStyle;
-import net.imagej.ui.swing.viewer.plot.utils.AwtLineStyles;
-import net.imagej.ui.swing.viewer.plot.utils.AwtMarkerStyles;
 import org.jfree.chart.renderer.AbstractRenderer;
-import org.jfree.chart.renderer.category.BarRenderer;
 import org.jfree.chart.renderer.category.LineAndShapeRenderer;
 import org.jfree.chart.renderer.xy.XYLineAndShapeRenderer;
 import org.scijava.ui.awt.AWTColors;
